+++ src/Core/hco_calc_mod.F90
@@ -1035,15 +1035,8 @@ CONTAINS
              TMPVAL = BaseDct%Dta%V3(tIDx)%Val(I,J,L)
           ENDIF
 
-          ! If it's a missing value, mask box as unused and set value to
-          ! zero
-#if defined( ESMF_ )
-          ! SDE 2017-01-07: Temporary kludge. MAPL ExtData sets missing
-          ! data to 1e15, but HEMCO uses a different value!
-          IF ( ( TMPVAL == HCO_MISSVAL ) .or. ( TMPVAL > 1.0e+14 ) ) THEN
-#else
+          ! If it's a missing value, mask box as unused and set value to zero
           IF ( TMPVAL == HCO_MISSVAL ) THEN
-#endif
              MASK(I,J,:)      = 0.0_hp
              OUTARR_3D(I,J,L) = 0.0_hp
 
@@ -3086,15 +3079,8 @@ END FUNCTION GetEmisLUnit
              TMPVAL = BaseDct%Dta%V3(tIDx)%Val(I,J,L)
           ENDIF
 
-          ! If it's a missing value, mask box as unused and set value to
-          ! zero
-#if defined( ESMF_ )
-          ! SDE 2017-01-07: Temporary kludge. MAPL ExtData sets missing
-          ! data to 1e15, but HEMCO uses a different value!
-          IF ( ( TMPVAL == HCO_MISSVAL ) .or. ( TMPVAL > 1.0e+14 ) ) THEN
-#else
+          ! If it's a missing value, mask box as unused and set value to zero
           IF ( TMPVAL == HCO_MISSVAL ) THEN
-#endif
              MASK(I,J,:)      = 0.0_hp
              OUTARR_3D(I,J,L) = 0.0_hp