+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased 3.6.0]
+### Added
+  - Added MAPL_ESMF compiler option for use with GCHP and GEOS
+
+### Changed
+  - Set HCO_MISSVAL to MAPL missing value (1e15) if using GCHP or GEOS
+  - Use fraction surface type inputs instead of ExtState%WLI
+
+### Fixed
+  - placeholder
+
+### Removed
+  - Removed old kludge for MAPL missing data if applying mask
+  - Removed ExtState field for water-land-ice index (WLI)
+
 ## [3.5.2] - 2022-11-29
 ### Added
   - Added sanitizer option for detecting memory leaks in HEMCO